Implements a graph isomorphism check for PimSubgraphs.
This is a big commit that needs to be cleaned up; however, this basic version is
mostly working. It's not thoroughly tested, but from my results, it seems to get
all comparisons correct.

I've also updated the printer to include subgraph "class" info, where
classes are determined by this comparison.