Makes sure that subgraph values are all in the same BasicBlock.
This implements this much-needed check.
I've run this new version to confirm that there's no crashes on the
benchmarks that I care about. However, it also looks like the results
may not have changed that much. Thus, either crossing BB boundaries was
rare in the past, or something's wrong with the implementation.